+++ jk_src/jk_lib/json/parse_test.c
@@ -18,7 +18,7 @@ int main(int argc, char **argv)
     jk_platform_console_utf8_enable();
 
     JkPlatformArenaVirtualRoot arena_root;
-    JkArena storage = jk_platform_arena_virtual_init(&arena_root, (size_t)1 << 35);
+    JkArena storage = jk_platform_arena_virtual_init(&arena_root, (int64_t)1 << 35);
 
     JkBuffer text = jk_platform_file_read_full(&storage, "../jk_src/jk_lib/json/parse_test.json");
