+++ jk_src/jk_lib/precision_test/precision_test.h
@@ -9,7 +9,7 @@ typedef struct JkPrecisionTestReference {
 } JkPrecisionTestReference;
 
 typedef struct JkPrecisionTestResult {
-    uint64_t diff_count;
+    int64_t diff_count;
     double diff_total;
     double diff_max;
     double input_at_diff_max;
@@ -18,18 +18,18 @@ typedef struct JkPrecisionTestResult {
 
 typedef struct JkPrecisionTest {
     double input;
-    uint64_t step_index;
-    uint64_t result_index;
-    uint64_t result_count;
+    int64_t step_index;
+    int64_t result_index;
+    int64_t result_count;
     JkPrecisionTestResult results[256];
 } JkPrecisionTest;
 
 JK_PUBLIC void jk_precision_test_check_reference(char *label,
         double (*function)(double),
-        uint32_t reference_count,
+        int64_t reference_count,
         JkPrecisionTestReference *references);
 
-JK_PUBLIC b32 jk_precision_test(JkPrecisionTest *t, double min, double max, uint64_t step_count);
+JK_PUBLIC b32 jk_precision_test(JkPrecisionTest *t, double min, double max, int64_t step_count);
 
 JK_PUBLIC void jk_precision_test_result(
         JkPrecisionTest *t, double reference, double value, char *label, ...);
